enemy collison with wall

+++ src/main/java/org/sfu/MotaGame/Bean/player/Enemy.java
@@ -42,10 +42,10 @@ public class Enemy extends Rectangle{
 	public double getY() {
 		return y;
 	}
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 	
@@ -85,19 +85,19 @@ public class Enemy extends Rectangle{
 	public Rectangle getBounds() {
 		return new Rectangle((int) x, (int) y, width, height);
 	}
-	
+
 	public Rectangle getBoundsTop() {
 		return new Rectangle((int) x + (width / 2) - ((width / 2) / 2), (int) y, (int) width / 2, (int) height / 2);
 	}
-	
+
 	public Rectangle getBoundsBottom() {
 		return new Rectangle((int) x + (width / 2) - ((width / 2) / 2), (int) y + (height / 2), (int) width / 2, (int) height / 2);
 	}
-	
+
 	public Rectangle getBoundsRight() {
 		return new Rectangle((int) x + width - 5, (int) y + 4, (int) 5, (int) height - 8);
 	}
-	
+
 	public Rectangle getBoundsLeft() {
 		return new Rectangle((int) x, (int) y + 4, 5, height - 8);
 	}
